reverting modifications for difficulty opcodes

+++ src/server/game/Entities/Creature/Creature.h
@@ -585,7 +585,7 @@ class Creature : public Unit, public GridObject<Creature>, public MapCreature
         bool LoadCreatureFromDB(uint32 guid, Map* map, bool addToMap = true);
         void SaveToDB();
                                                             // overriden in Pet
-        virtual void SaveToDB(uint32 mapid, uint8 spawnMask, uint32 phaseMask);
+        virtual void SaveToDB(uint32 mapid, uint32 spawnMask, uint32 phaseMask);
         virtual void DeleteFromDB();                        // overriden in Pet
 
         Loot loot;
